# docs/OpenApi/letterbook-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Letterbook APIs",
    "description": "Letterbook's first party API",
    "contact": {
      "url": "https://letterbook.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/lb/v1/reports/{selfId}/report": {
      "post": {
        "tags": [
          "Reports"
        ],
        "summary": "Create",
        "description": "Open a new moderation report",
        "parameters": [
          {
            "name": "selfId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProfileId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberModerationReportDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberModerationReportDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MemberModerationReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/{selfId}/report/{reportId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get",
        "description": "Get a new moderation report by ID",
        "parameters": [
          {
            "name": "selfId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProfileId"
            }
          },
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModerationReportId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/moderator/report": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "List",
        "description": "Filter and list moderation reports",
        "parameters": [
          {
            "name": "subjectId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProfileId"
            }
          },
          {
            "name": "reporterId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProfileId"
            }
          },
          {
            "name": "moderatorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeClosed",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FullModerationReportDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FullModerationReportDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FullModerationReportDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/moderator/report/{reportId}": {
      "put": {
        "tags": [
          "Reports"
        ],
        "summary": "Update",
        "description": "Update a moderation report",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModerationReportId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullModerationReportDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FullModerationReportDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FullModerationReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/moderator/report/{reportId}/remark": {
      "post": {
        "tags": [
          "Reports"
        ],
        "summary": "Remark",
        "description": "Add a remark to a moderation report",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModerationReportId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRemarkDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRemarkDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRemarkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/moderator/report/{reportId}/assign/{accountId}": {
      "put": {
        "tags": [
          "Reports"
        ],
        "summary": "Assign",
        "description": "Assign or unassign a moderator",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModerationReportId"
            }
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "unassign",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/lb/v1/reports/moderator/report/{reportId}/close": {
      "put": {
        "tags": [
          "Reports"
        ],
        "summary": "Close",
        "description": "Close or reopen a moderation report",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModerationReportId"
            }
          },
          {
            "name": "reopen",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberModerationReportDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FullModerationReportDto": {
        "required": [
          "context",
          "summary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ModerationReportId"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationPolicyId"
            },
            "nullable": true
          },
          "reporter": {
            "$ref": "#/components/schemas/ProfileId"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileId"
            },
            "nullable": true
          },
          "relatedPosts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostId"
            },
            "nullable": true
          },
          "forwarded": {
            "type": "array",
            "items": {
              "pattern": "https://\\w+\\.\\w+/\\w*",
              "type": "string",
              "example": "https://example.com/fgfrff01jnp851pm72jmn9v44"
            },
            "nullable": true
          },
          "fediId": {
            "pattern": "https://\\w+\\.\\w+/\\w*",
            "type": "string",
            "nullable": true,
            "example": "https://example.com/fgfrff01jnp851pm72jmn9v44"
          },
          "context": {
            "$ref": "#/components/schemas/ThreadId"
          },
          "moderators": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "closed": {
            "type": "string",
            "format": "date-time"
          },
          "remarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationRemarkDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MemberModerationReportDto": {
        "required": [
          "summary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ModerationReportId"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationPolicyId"
            },
            "nullable": true
          },
          "reporter": {
            "$ref": "#/components/schemas/ProfileId"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileId"
            },
            "nullable": true
          },
          "relatedPosts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostId"
            },
            "nullable": true
          },
          "forwarded": {
            "type": "array",
            "items": {
              "pattern": "https://\\w+\\.\\w+/\\w*",
              "type": "string",
              "example": "https://example.com/fgfrff01jnp851pm72jmn9v44"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModerationPolicyId": {
        "type": "integer",
        "additionalProperties": false,
        "example": "0"
      },
...
      "ModerationRemarkDto": {
        "required": [
          "author",
          "report",
          "text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ModerationRemarkId"
          },
          "report": {
            "$ref": "#/components/schemas/ModerationReportId"
          },
          "author": {
            "type": "string",
            "format": "uuid"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModerationRemarkId": {
        "type": "string",
        "additionalProperties": false,
        "example": "0000000000000000000000000"
      },
...
      "ModerationReportId": {
        "type": "string",
        "additionalProperties": false,
        "example": "0000000000000000000000000"
      },
...
      "PostId": {
        "type": "string",
        "additionalProperties": false,
        "example": "0000000000000000000000000"
      },
...
      "ProfileId": {
        "type": "string",
        "additionalProperties": false,
        "example": "0000000000000000000000000"
      },
...
      "ThreadId": {
        "type": "string",
        "additionalProperties": false,
        "example": "0000000000000000000000000"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter the Authorization header, including the Bearer scheme, like so: `Bearer <JWT>`",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}